move students to loans
